requirements.txt
aiobotocore==2.9.0
aiofiles==23.2.1
aiohttp==3.9.1
azure-cognitiveservices-speech==1.38.0
daily-python==0.9.1
fastapi==0.108.0
fastembed==0.2.7
litellm==1.40.20
numpy==1.26.1
openai>=1.10.0
pydantic==2.5.3
pydub==0.25.1
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
tiktoken>=0.6.0
torchaudio==2.0.1
twilio==8.9.0
uvicorn==0.22.0
websockets==10.4
onnxruntime>=1.16.3
scipy==1.11.4
uvloop==0.19.0
tokenizers==0.15.2
huggingface-hub==0.20.1
semantic-router==0.0.46
sentence-transformers==3.0.1
optimum[onnxruntime]